Adding RSA_Client_App and rsa memory mgmt library
  - Adding a simple server which will be used to (de)serialize and Rx/Tx data
regarding memory requested by a Dom0 client who's managing a DomU
  - Adding an interface that both the libbdvmi example app and the
server can use to request more memory, or send memory
  - Adding a custom buffer and data (de)serialization library which
works in pair with the RSA_Mem_Data struct
  - Adding an example client which proves that the server/client
interface works

This application is going to be the foundation of the memory server.
When the libbdvmi rsa_app requests more memory from a remote host for a
DomU, this server will be responsible for calling malloc and maintaining
a record of which DomU is using which page of memory.

The client application is a just a test app to assure the server and the
network interface works.  It also is a test app to assure that the data
(de)serialization interface works.